Take out `setwd`
You can assume that you're already working in this directory anytime you work on this Project-- the default is to run this project file from this working directory. Setting the directory like this will cause the code to not run to other people who download the project, so it's better to take out.