Take out adding mice IDs
This was a *very* nice solution to the limitations of the original dataset templates (the Gates one). However, I think it's reasonable for us to tell the PI that this template *must* have mouse IDs with each row of data. I've added that information into *our* template for laser data, so I've edited the function to read mouse ID directly from the data.