Filter to only Steady State
So, it looks like the template only records Cmax and Trough for the observations where "Dosing" is "Steady_State". The PI might someday want to work in the single dose values, but since we aren't using them down the line right now, I think it's okay that we filter them out early to get the data cleaned up and let the PI know that the current prototype isn't adding in that information.