Add `our_templates` folder, with our laser template
I'm mocking up templates (with fake data) that add any columns we think are necessary (to let us get from raw data to the summarized efficacy dataset), and with numbers that will get us to the summarized data that the other groups are using as their inputs. We can share these templates with the research group and let the know that this is the template they would need to put their data in to use this app. For the laser data, the key additions are: (1) MouseID must be recorded (unique for every research animal) (2) added a column for drug dose and (3) added a column for dose frequency.