Add a summary function for laser data
This function works from the output of Lizette's function for cleaning the laser data (with revisions based on our changes to the template for raw laser data). This gets the data in the format we'll need for the summarized efficacy data frame that the other groups are starting from for most of their functions.