Added our template for TIssue Std PK
This version is very similar to the previous edited "Gates" template. Key changes are (1) required to give a unique mouse ID for every mouse, (2) column added for drug dose, (3) column added for dose frequency. These will provide the information needed to summarize the data to get to the format of the efficacy summary table.