Take out code adding mouse ID
As with the laser file, the original code here was a very nice solution to the limitations of the original "Gates" template data. I think it is reasonable for us to establish a template that requires the research group to enter unique mouse IDs for every different animal. Otherwise, it would be very easy for the code to get the mouse number wrong (for example, if the researcher accidently skips adding one row for one mouse somewhere along the way).