Use `mutate_at` instead of `mutate`
This is a very minor change, but you can save some code by using `mutate_at` when you need to make the same change to more than one column. Here, it doesn't save much time, but if you needed to change a bunch of columns to numeric, for example, it can be very helpful.