Filter to steady state values
Groups 2 and 3 are focusing on the Cmax and Trough values, and it looks like those are only collected for the "Steady_State" dosing. We can add a filter to get the cleaned data down to just those rows for right now. If they researchers want to change the analysis to also include single dose observations, they could take this filtering step out in the future