Use information in `Group` column to label baseline
Previously, we were marking anything with a missing drug as "Baseline". It might be more robust to use the information in the `Group` column ("PreRX" for animals before treatment began) and use logic based on that (I put an `ifelse` call in here) to relabel the drug observations that should be "Baseline".