Add in a function to create efficacy summary
This function pulls together all the other summarized datasets to create the efficacy summary dataframe that groups 2 and 3 use as the start for most of their work, connecting through from the raw data in our example templates. Also I changed a few pieces of helper functions as needed to get the summarizing function to work.